# docs/api-contracts.yaml
openapi: 3.0.3
info:
  title: EduPro Suite API
  description: Comprehensive Educational Management System API for Bangladesh
  version: 1.0.0
  contact:
    name: EduPro Suite Development Team
    email: support@edupro-suite.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://edupro-suite.com/api
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items: {}
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
paths:
...
  /admin/finance/payments:
    get:
      tags:
        - Finance Management
      summary: Get payments
      responses:
        '200':
          description: Payments retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse'

    post:
      tags:
        - Finance Management
      summary: Record payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                studentId:
                  type: string
                amount:
                  type: number
                paymentMethod:
                  type: string
                  enum: [CASH, CARD, MOBILE_BANKING, BANK_TRANSFER]
                transactionId:
                  type: string
      responses:
        '201':
          description: Payment recorded

  # Assignment Management
...